pyproject.toml
[build-system]
requires = ["setuptools >= 61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "credit_fraud"
version = "1.0.0"
authors = [
    {name = "Eduardo B. Gonçalves", email = "eduardo.goncalves@f1rst.com.br"},
]
description = "Credit Fraud Detection MLOps End-to-End Project."
readme = "README.md"
requires-python = ">=3.11"
keywords = ["credit", "fraud", "mlops", "aws"]
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "requests",
    "scikit-learn==1.4.2",
    "pandas==2.2.2",
    "xgboost==1.7.6",
    "fsspec==2024.5.0",
    "s3fs~=2024.5.0",
    "sagemaker==2.219.0",
    "boto3~=1.34",
    "python-dotenv==1.0.1",
    "toml==0.10.2",
    "pymysql==1.1.1",
    "pytest==8.2.2",
    "pytest-mock==3.14.0",
    "pytest-cov==5.0.0",
    "mlflow~=2.14",
    "sagemaker-mlflow"
]

[tool.setuptools]
packages = [
    "credit_fraud",
    "credit_fraud.pipeline",
    "credit_fraud.pipeline.steps",
    "credit_fraud.utils"
]

[project.gui-scripts]
cf-run = "credit_fraud:run"

[tool.pytest.ini_options]
addopts = """
--cov-config pyproject.toml \
--cov-report xml:coverage.xml \
--cov-report term \
--cov credit_fraud \
--color yes \
--cov-fail-under 10 \
-ra \
-W ignore"""
testpaths = [
    "tests"
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pycodestyle]
max-doc-length = 100

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint]
select = [
    "E",
    "F"
]
ignore = [
    "F403"
]

[tool.ruff]
exclude = [
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".pytest_cache",
    ".ruff_cache",
    ".svn",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "docs"
]

line-length = 88
indent-width = 4
